# New: drop gzip support, improve performance.