@steve-chavez 2021-05-21
> Though, it seems easier to explain to people: "We don't support compression - add a proxy to do it", instead of "yeah, we have compression, but it's not very good and also easily misused...".

Yeah, I was thinking we could have a config option like `gzip_enabled = false` by default. But I don't think keeping our gzip is worth it if it impacts performance in surprising ways. It's a better job for the proxy since it provides more options as well.

So +1 for removing it as well.