@aljungberg 2021-05-21
> But without the change I could still have a nginx proxy before postgrest to handle the gzip. If that one removes the header, I would get the same performance as after dropping gzip support in postgrest, correct?

> Though, it seems easier to explain to people: "We don't support compression - add a proxy to do it", instead of "yeah, we have compression, but it's not very good and also easily misused...".

Right. And nothing is actually lost by removing this feature (unless you're running a bare Postgrest). Previously, you could control compression in your proxy (by changing a header). After this change, you can control compression in your proxy (by enabling compression).

And your proxy is *better* at it (faster, more configurable, less surprising). I think the rationale for dropping this feature in Postgrest is simply that it's too easy to accidentally enable it and you basically never should. I only discovered I had it accidentally enabled by load testing.

Maybe there's some theoretical use case when you pay for bandwidth between Postgrest and your proxy or they're in different DCs. I think that falls under "don't solve problems you don't have".

(Also there must be *some* overhead even when it's disabled via accept encoding since it's one more layer of middleware and the headers have to be checked.)