@steve-chavez 2021-05-21
Also, if users want to gzip without proxy, maybe they could do it inside PostgreSQL as well: https://github.com/pramsey/pgsql-gzip (I haven't tried it, but looks feasible)