@steve-chavez 2021-05-30
So I think this one should be good to merge. Though we don't edit the CHANGELOG anymore, I think the change is harmless. I can squash the commits and edit the commit message to go according to our format on https://github.com/PostgREST/postgrest/pull/1805#issuecomment-817350164.